API - Securing team deletion

+++ api/src/main/java/com/gberard/tournament/domain/service/TeamServiceImpl.java
@@ -1,7 +1,9 @@
 package com.gberard.tournament.domain.service;
 
 import com.gberard.tournament.domain.model.Team;
+import com.gberard.tournament.domain.exception.TeamInUseException;
 import com.gberard.tournament.domain.port.input.TeamService;
+import com.gberard.tournament.domain.port.output.GameRepository;
 import com.gberard.tournament.domain.port.output.TeamRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -15,6 +17,9 @@ public class TeamServiceImpl implements TeamService {
     @Autowired
     public TeamRepository teamRepository;
 
+    @Autowired
+    public GameRepository gameRepository;
+
     @Override
     public Team create(Team player) {
         return teamRepository.save(player);
@@ -27,6 +32,10 @@ public class TeamServiceImpl implements TeamService {
 
     @Override
     public boolean delete(Team player) {
+        if (gameRepository.existsByTeamId(player.id())) {
+            throw new TeamInUseException(player.id());
+        }
+
         teamRepository.deleteById(player.id());
         return true;
     }
